bug fixed of time counting in linux. And other include problems solved.

+++ src/main/main_omp/ModelMain.h
@@ -17,6 +17,7 @@
 #include "clsRasterData.h"
 #include "mongoc.h"
 #include "ModuleFactory.h"
+#include "utils.h"
 
 using namespace std;
 
@@ -77,7 +78,7 @@ public:
     string GetModuleID(int i) { return m_factory->GetModuleID(i); }
 
     //! Get module execute time by index in ModuleFactory
-    int GetModuleExecuteTime(int i) { return m_executeTime[i]; }
+    float GetModuleExecuteTime(int i) { return m_executeTime[i]; }
 
     //! Get time consuming of read files
     int GetReadDataTime() { return m_readFileTime; }
@@ -145,7 +146,7 @@ private:
     //! Whole simulation scale modules index list
     vector<int> m_overallModules;
     //! Execute time list of each module
-    vector<int> m_executeTime;
+    vector<double> m_executeTime;
     //! Time consuming for read files
     int m_readFileTime;
     //! Layering method
